examples/fastapi-mongo/pyproject.toml
[tool.poetry]
name = "fastapi-mongo"
version = "0.1.0"
description = "Hello World FastAPI example"
authors = ["DeployPlex Team <hi@deployplex.com>"]
readme = "README.md"
packages = [{include = "fastapi_mongo"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
python-dotenv = "^1.0.0"
motor = "^3.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
